chore(toc): refactored toc
- added fetch Name to TMID
- added isOfficial() method to ThingModel for TMID parsing
- moved Insert() function to model package as a method of TOC
- Insert() relies now solely on CatalogThingModel argument
- removed any path computation from TOC, relying on TMID as URL
- moved reading/walking/writing filesystem to fs remote
