+++ internal/model/model.go
@@ -1,5 +1,7 @@
 package model
 
+import "github.com/web-of-things-open-source/tm-catalog-cli/internal"
+
 // ThingModel is a model for unmarshalling a Thing Model to be
 // imported. It contains only the fields required to be accepted into
 // the catalog.
@@ -8,6 +10,12 @@ type ThingModel struct {
 	Mpn          string             `json:"schema:mpn" validate:"required"`
 	Author       SchemaAuthor       `json:"schema:author" validate:"required"`
 	Version      Version            `json:"version"`
+}
+
+func (tm *ThingModel) IsOfficial() bool {
+	compareMan := internal.ToTrimmedLower(tm.Manufacturer.Name)
+	compareAuthor := internal.ToTrimmedLower(tm.Author.Name)
+	return compareMan == compareAuthor
 }
 
 type SchemaAuthor struct {
@@ -22,7 +30,8 @@ type Version struct {
 }
 
 type ExtendedFields struct {
-	Links       `json:"links"`
+	Links `json:"links"`
+	// TODO: why is ID field not in ThingModel? It is also needed for importing to be moved to "original"
 	ID          string `json:"id,omitempty"`
 	Description string `json:"description"`
 }
